+++ quantify_scheduler/device_under_test/transmon_element.py
@@ -530,7 +530,7 @@ class BasicTransmonElement(DeviceElement):
     def generate_device_config(self) -> DeviceCompilationConfig:
         """
         Generates a valid device config for the quantify-scheduler making use of the
-        :func:`~.circuit_to_device.compile_circuit_to_device` function.
+        :func:`~.circuit_to_device._compile_circuit_to_device` function.
 
         This enables the settings of this qubit to be used in isolation.
 
@@ -540,7 +540,7 @@ class BasicTransmonElement(DeviceElement):
         """
         cfg_dict = {
             "backend": "quantify_scheduler.backends"
-            ".circuit_to_device.compile_circuit_to_device",
+            ".circuit_to_device._compile_circuit_to_device",
             "elements": self._generate_config(),
             "clocks": {
                 f"{self.name}.01": self.clock_freqs.f01(),