+++ quantify_scheduler/schemas/examples/device_example_cfgs.py
@@ -3,7 +3,7 @@
 
 example_transmon_cfg = {
     "backend": "quantify_scheduler.backends.circuit_to_device"
-    + ".compile_circuit_to_device",
+    + "._compile_circuit_to_device",
     "clocks": {
         "q0.01": 6020000000.0,
         "q0.ro": 7040000000.0,