+++ tests/scheduler/backends/test_circuit_to_device.py
@@ -12,7 +12,7 @@ from quantify_scheduler.backends.circuit_to_device import (
     ConfigKeyError,
     DeviceCompilationConfig,
     OperationCompilationConfig,
-    compile_circuit_to_device,
+    _compile_circuit_to_device,
     set_pulse_and_acquisition_clock,
     _clocks_compatible,
     _valid_clock_in_schedule,
@@ -77,9 +77,7 @@ def test_compile_all_gates_example_transmon_cfg():
     assert len(sched.schedulables) == 18
 
     # test that all these operations compile correctly.
-    _ = compile_circuit_to_device(
-        sched, device_cfg=example_transmon_cfg, keep_original_schedule=False
-    )
+    _ = _compile_circuit_to_device(sched, device_cfg=example_transmon_cfg)
 
 
 def test_compile_all_gates_basic_transmon(mock_setup_basic_transmon):
@@ -114,10 +112,9 @@ def test_compile_all_gates_basic_transmon(mock_setup_basic_transmon):
 
     # test that all these operations compile correctly.
     quantum_device = mock_setup_basic_transmon["quantum_device"]
-    _ = compile_circuit_to_device(
+    _ = _compile_circuit_to_device(
         sched,
         device_cfg=quantum_device.generate_device_config(),
-        keep_original_schedule=False,
     )
 
 
@@ -144,10 +141,8 @@ def test_compile_asymmetric_gate(mock_setup_basic_transmon):
     quantum_device = mock_setup_basic_transmon["quantum_device"]
 
     with pytest.raises(ConfigKeyError):
-        _ = compile_circuit_to_device(
-            sched,
-            device_cfg=quantum_device.generate_device_config(),
-            keep_original_schedule=False,
+        _ = _compile_circuit_to_device(
+            sched, device_cfg=quantum_device.generate_device_config()
         )
 
 
@@ -157,9 +152,7 @@ def test_measurement_compile():
     sched.add(Measure("q0", acq_index=1))
     sched.add(Measure("q1", acq_index=2))  # acq_channel should be 1
     sched.add(Measure("q0", "q1", acq_index=2))
-    new_dev_sched = compile_circuit_to_device(
-        sched, device_cfg=example_transmon_cfg, keep_original_schedule=False
-    )
+    new_dev_sched = _compile_circuit_to_device(sched, device_cfg=example_transmon_cfg)
 
     operation_keys_list = list(new_dev_sched.operations.keys())
 
@@ -205,9 +198,7 @@ def test_only_add_clocks_used(operations: List[Operation], clocks_used: List[str
     sched = Schedule("Test schedule")
     for operation in operations:
         sched.add(operation)
-    dev_sched = compile_circuit_to_device(
-        sched, device_cfg=example_transmon_cfg, keep_original_schedule=False
-    )
+    dev_sched = _compile_circuit_to_device(sched, device_cfg=example_transmon_cfg)
     checked_dev_sched = set_pulse_and_acquisition_clock(
         dev_sched, device_cfg=example_transmon_cfg
     )
@@ -250,9 +241,7 @@ def test_multiply_defined_clock_freq_raises(
     sched.add_resource(ClockResource(name="q0.01", freq=clock_freq_schedule))
     operation = X("q0")
     sched.add(operation)
-    dev_sched = compile_circuit_to_device(
-        schedule=sched, device_cfg=device_cfg, keep_original_schedule=False
-    )
+    dev_sched = _compile_circuit_to_device(schedule=sched, device_cfg=device_cfg)
 
     with pytest.warns(RuntimeWarning) as warning:
         compiled_sched = set_pulse_and_acquisition_clock(
@@ -293,7 +282,7 @@ def test_set_device_cfg_clock(
 def test_clock_not_defined_raises():
     simple_config = DeviceCompilationConfig(
         backend="quantify_scheduler.backends.circuit_to_device"
-        + ".compile_circuit_to_device",
+        + "._compile_circuit_to_device",
         clocks={
             "q0.01": 6020000000.0,
         },
@@ -329,9 +318,7 @@ def test_clock_not_defined_raises():
     sched = Schedule("Test schedule")
     operation = Measure("q0", acq_protocol="Trace")
     sched.add(operation)
-    dev_sched = compile_circuit_to_device(
-        sched, device_cfg=simple_config, keep_original_schedule=False
-    )
+    dev_sched = _compile_circuit_to_device(sched, device_cfg=simple_config)
     with pytest.raises(ValueError) as error:
         _ = set_pulse_and_acquisition_clock(dev_sched, device_cfg=simple_config)
 
@@ -347,56 +334,44 @@ def test_reset_operations_compile():
     sched = Schedule("Test schedule")
     sched.add(Reset("q0"))
     sched.add(Reset("q0", "q1"))
-    _ = compile_circuit_to_device(
-        sched, device_cfg=example_transmon_cfg, keep_original_schedule=False
-    )
+    _ = _compile_circuit_to_device(sched, device_cfg=example_transmon_cfg)
 
 
 def test_qubit_not_in_config_raises():
     sched = Schedule("Test schedule")
     sched.add(Rxy(90, 0, qubit="q20"))
     with pytest.raises(ConfigKeyError):
-        _ = compile_circuit_to_device(
-            sched, device_cfg=example_transmon_cfg, keep_original_schedule=False
-        )
+        _ = _compile_circuit_to_device(sched, device_cfg=example_transmon_cfg)
 
     sched = Schedule("Test schedule")
     sched.add(Reset("q2", "q5", "q3"))
     with pytest.raises(ConfigKeyError):
-        _ = compile_circuit_to_device(
-            sched, device_cfg=example_transmon_cfg, keep_original_schedule=False
-        )
+        _ = _compile_circuit_to_device(sched, device_cfg=example_transmon_cfg)
 
     sched = Schedule("Test schedule")
     sched.add(Reset("q0", "q5"))
     with pytest.raises(ConfigKeyError):
-        _ = compile_circuit_to_device(
-            sched, device_cfg=example_transmon_cfg, keep_original_schedule=False
-        )
+        _ = _compile_circuit_to_device(sched, device_cfg=example_transmon_cfg)
 
 
 def test_edge_not_in_config_raises():
     sched = Schedule("Test schedule")
     sched.add(CZ("q0", "q3"))
     with pytest.raises(ConfigKeyError):
-        _ = compile_circuit_to_device(
-            sched, device_cfg=example_transmon_cfg, keep_original_schedule=False
-        )
+        _ = _compile_circuit_to_device(sched, device_cfg=example_transmon_cfg)
 
 
 def test_operation_not_in_config_raises():
     sched = Schedule("Test schedule")
     sched.add(CNOT("q0", "q1"))
     with pytest.raises(ConfigKeyError):
-        _ = compile_circuit_to_device(
-            sched, device_cfg=example_transmon_cfg, keep_original_schedule=False
-        )
+        _ = _compile_circuit_to_device(sched, device_cfg=example_transmon_cfg)
 
 
 def test_compile_schedule_with_trace_acq_protocol():
     simple_config = DeviceCompilationConfig(
         backend="quantify_scheduler.backends.circuit_to_device"
-        + ".compile_circuit_to_device",
+        + "._compile_circuit_to_device",
         clocks={
             "q0.01": 6020000000.0,
             "q0.ro": 7040000000.0,
@@ -430,15 +405,13 @@ def test_compile_schedule_with_trace_acq_protocol():
     )
     sched = Schedule("Test schedule")
     sched.add(Measure("q0", acq_protocol="Trace"))
-    _ = compile_circuit_to_device(
-        sched, device_cfg=simple_config, keep_original_schedule=False
-    )
+    _ = _compile_circuit_to_device(sched, device_cfg=simple_config)
 
 
 def test_compile_schedule_with_invalid_pulse_type_raises():
     simple_config = DeviceCompilationConfig(
         backend="quantify_scheduler.backends.circuit_to_device"
-        + ".compile_circuit_to_device",
+        + "._compile_circuit_to_device",
         clocks={
             "q0.01": 6020000000.0,
             "q0.ro": 7040000000.0,
@@ -473,15 +446,13 @@ def test_compile_schedule_with_invalid_pulse_type_raises():
     sched = Schedule("Test schedule")
     sched.add(Measure("q0", acq_protocol="Trace"))
     with pytest.raises(NotImplementedError):
-        _ = compile_circuit_to_device(
-            sched, device_cfg=simple_config, keep_original_schedule=False
-        )
+        _ = _compile_circuit_to_device(sched, device_cfg=simple_config)
 
 
 def test_operation_not_in_config_raises_custom():
     simple_config = DeviceCompilationConfig(
         backend="quantify_scheduler.backends.circuit_to_device"
-        + ".compile_circuit_to_device",
+        + "._compile_circuit_to_device",
         clocks={
             "q0.01": 6020000000.0,
             "q0.ro": 7040000000.0,
@@ -495,29 +466,23 @@ def test_operation_not_in_config_raises_custom():
     sched = Schedule("Test missing single q op")
     sched.add(Reset("q0"))
     with pytest.raises(ConfigKeyError):
-        _ = compile_circuit_to_device(
-            sched, device_cfg=simple_config, keep_original_schedule=False
-        )
+        _ = _compile_circuit_to_device(sched, device_cfg=simple_config)
 
     sched = Schedule("Test schedule mux missing op")
     sched.add(Reset("q0", "q1", "q2"))
     with pytest.raises(ConfigKeyError):
-        _ = compile_circuit_to_device(
-            sched, device_cfg=simple_config, keep_original_schedule=False
-        )
+        _ = _compile_circuit_to_device(sched, device_cfg=simple_config)
 
     sched = Schedule("Test missing 2Q op")
     sched.add(CZ("q0", "q1"))
     with pytest.raises(ConfigKeyError):
-        _ = compile_circuit_to_device(
-            sched, device_cfg=simple_config, keep_original_schedule=False
-        )
+        _ = _compile_circuit_to_device(sched, device_cfg=simple_config)
 
 
 def test_config_with_callables():
     simple_config = DeviceCompilationConfig(
         backend="quantify_scheduler.backends.circuit_to_device"
-        + ".compile_circuit_to_device",
+        + "._compile_circuit_to_device",
         clocks={
             "q0.01": 6020000000.0,
             "q0.ro": 7040000000.0,
@@ -555,9 +520,7 @@ def test_config_with_callables():
 
     sched = Schedule("Test callable op")
     sched.add(Reset("q0", "q1"))
-    _ = compile_circuit_to_device(
-        sched, device_cfg=simple_config, keep_original_schedule=False
-    )
+    _ = _compile_circuit_to_device(sched, device_cfg=simple_config)
 
 
 def test_config_validation():
@@ -693,10 +656,8 @@ def test_set_reference_magnitude(mock_setup_basic_transmon):
 
     # test that all these operations compile correctly.
     quantum_device = mock_setup_basic_transmon["quantum_device"]
-    compiled_schedule = compile_circuit_to_device(
-        sched,
-        device_cfg=quantum_device.generate_device_config(),
-        keep_original_schedule=False,
+    compiled_schedule = _compile_circuit_to_device(
+        sched, device_cfg=quantum_device.generate_device_config()
     )
 
     operations_dict_with_repr_keys = {