+++ tests/scheduler/backends/test_qblox_backend.py
@@ -52,7 +52,7 @@ from quantify_scheduler.backends.qblox.instrument_compilers import (
 from quantify_scheduler.backends.qblox.qasm_program import QASMProgram
 from quantify_scheduler.backends.types import qblox as types
 from quantify_scheduler.backends.types.qblox import BasebandModuleSettings
-from quantify_scheduler.compilation import determine_absolute_timing
+from quantify_scheduler.compilation import _determine_absolute_timing
 from quantify_scheduler.device_under_test.quantum_device import QuantumDevice
 from quantify_scheduler.helpers.collections import (
     find_all_port_clock_combinations,
@@ -1624,9 +1624,7 @@ def test_container_prepare_no_lo(
 
 
 def test_container_add_from_type(pulse_only_schedule, hardware_cfg_pulsar):
-    pulse_only_schedule = determine_absolute_timing(
-        pulse_only_schedule, keep_original_schedule=False
-    )
+    pulse_only_schedule = _determine_absolute_timing(pulse_only_schedule)
     container = compiler_container.CompilerContainer(pulse_only_schedule)
     container.add_instrument_compiler("qcm0", QcmModule, hardware_cfg_pulsar["qcm0"])
     assert "qcm0" in container.instrument_compilers
@@ -1634,9 +1632,7 @@ def test_container_add_from_type(pulse_only_schedule, hardware_cfg_pulsar):
 
 
 def test_container_add_from_str(pulse_only_schedule, hardware_cfg_pulsar):
-    pulse_only_schedule = determine_absolute_timing(
-        pulse_only_schedule, keep_original_schedule=False
-    )
+    pulse_only_schedule = _determine_absolute_timing(pulse_only_schedule)
     container = compiler_container.CompilerContainer(pulse_only_schedule)
     container.add_instrument_compiler("qcm0", "Pulsar_QCM", hardware_cfg_pulsar["qcm0"])
     assert "qcm0" in container.instrument_compilers
@@ -1644,9 +1640,7 @@ def test_container_add_from_str(pulse_only_schedule, hardware_cfg_pulsar):
 
 
 def test_container_add_from_path(pulse_only_schedule, hardware_cfg_pulsar):
-    pulse_only_schedule = determine_absolute_timing(
-        pulse_only_schedule, keep_original_schedule=False
-    )
+    pulse_only_schedule = _determine_absolute_timing(pulse_only_schedule)
     container = compiler_container.CompilerContainer(pulse_only_schedule)
     container.add_instrument_compiler(
         "qcm0",
@@ -1660,9 +1654,7 @@ def test_container_add_from_path(pulse_only_schedule, hardware_cfg_pulsar):
 def test_from_mapping(
     pulse_only_schedule, compile_config_basic_transmon_qblox_hardware
 ):
-    pulse_only_schedule = determine_absolute_timing(
-        pulse_only_schedule, keep_original_schedule=False
-    )
+    pulse_only_schedule = _determine_absolute_timing(pulse_only_schedule)
     hardware_cfg = generate_hardware_config(
         compilation_config=compile_config_basic_transmon_qblox_hardware
     )
@@ -1693,9 +1685,7 @@ def test_real_mode_container(
     hardware_cfg_pulsar_qcm_real_mode,
     mock_setup_basic_transmon,
 ):
-    real_square_pulse_schedule = determine_absolute_timing(
-        real_square_pulse_schedule, keep_original_schedule=False
-    )
+    real_square_pulse_schedule = _determine_absolute_timing(real_square_pulse_schedule)
     container = compiler_container.CompilerContainer.from_hardware_cfg(
         real_square_pulse_schedule, hardware_cfg_pulsar_qcm_real_mode
     )
@@ -2680,9 +2670,7 @@ def test_extract_settings_from_mapping(hardware_cfg_rf, hardware_cfg_pulsar):
 def test_cluster_settings(
     pulse_only_schedule, compile_config_basic_transmon_qblox_hardware
 ):
-    pulse_only_schedule = determine_absolute_timing(
-        pulse_only_schedule, keep_original_schedule=False
-    )
+    pulse_only_schedule = _determine_absolute_timing(pulse_only_schedule)
     hardware_cfg = generate_hardware_config(
         compilation_config=compile_config_basic_transmon_qblox_hardware
     )